(QENG-3890) Webhook proxy
This sets up a node to proxy webhook requests from external sources
though to specific URLs internally.

This enables webhooks to all of our Jenkins instances that require
maintenance windows, to jenkins-cinext, and to the SysOps Jenkins.

This allows event notification from external sources. For example,
GitHub can notify Jenkins when a branch is updated, causing a job to run
immediately rather than needing to poll.