(OPS-8422, OPS-8900) GitHub mirror
This creates a mirror of the entire puppetlabs organization in GitHub
(or at least as much of it as the plops-githubmirror user can access).

The mirror is updated with webhooks, so it will stay very fresh.
Additionally, repos are checked for updates at least once a day.

This is running on `mirror-github-prod-1`.